feat: Add DTOs for daily tasks, daily plans, monthly reports, and note tasks
- Created CreateDailyTaskDto for task creation.
- Added DailyPlanDto for daily plan representation.
- Introduced DailyTaskDto for detailed task information.
- Implemented MonthlyReportDto for monthly task statistics.
- Developed NoteTaskDto for note-related tasks.
- Added ReorderDailyTasksDto for task reordering functionality.

feat: Implement model classes for daily tasks, daily plans, and note tasks

- Created DailyPlan model with JPA annotations.
- Implemented DailyTask model with unique task ID and completion tracking.
- Added DailyTaskHistory model for task history tracking.
- Developed NoteTask model with support for subtasks.

feat: Create repositories for data access

- Added DailyPlanRepository for daily plan operations.
- Implemented DailyTaskRepository for daily task management.
- Created DailyTaskHistoryRepository for task history queries.
- Developed NoteTaskRepository for note task operations.
- Added TaskRepository for general task management.

feat: Implement services for business logic

- Created DailyTaskService for managing daily tasks and plans.
- Implemented NoteTaskService for handling note tasks.
- Developed TaskService for general task operations.

+++ src/main/java/com/mywebcompanion/backendspring/service/LabelService.java
@@ -22,7 +22,7 @@ public class LabelService {
 
     public List<LabelDto> getAllLabelsByUserId(String clerkId) {
         User user = userService.findByClerkId(clerkId);
-        return labelRepository.findByUserIdOrderByNameAsc(user.getId())
+        return labelRepository.findByUserIdOrderByNameAsc(user.getId().intValue())
                 .stream()
                 .map(this::convertToDto)
                 .collect(Collectors.toList());
@@ -32,7 +32,7 @@ public class LabelService {
         User user = userService.findByClerkId(clerkId);
 
         // Vérifier si le label existe déjà
-        if (labelRepository.findByNameAndUserId(name, user.getId()).isPresent()) {
+        if (labelRepository.findByNameAndUserId(name, user.getId().intValue()).isPresent()) {
             throw new RuntimeException("Un label avec ce nom existe déjà");
         }
 
@@ -60,7 +60,7 @@ public class LabelService {
         }
 
         // Vérifier si un autre label avec ce nom existe
-        labelRepository.findByNameAndUserId(name, user.getId())
+        labelRepository.findByNameAndUserId(name, user.getId().intValue())
                 .ifPresent(existingLabel -> {
                     if (!existingLabel.getId().equals(labelId)) {
                         throw new RuntimeException("Un label avec ce nom existe déjà");
